Move HTTP build cache config build operation test into its project.
