+++ subprojects/core/src/integTest/groovy/org/gradle/caching/internal/FinalizeBuildCacheConfigurationBuildOperationIntegrationTest.groovy
@@ -16,36 +16,15 @@
 
 package org.gradle.caching.internal
 
+import org.gradle.caching.local.DirectoryBuildCache
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.BuildOperationsFixture
-import org.gradle.test.fixtures.server.http.HttpBuildCache
 import org.junit.Rule
-import spock.lang.Unroll
 
 class FinalizeBuildCacheConfigurationBuildOperationIntegrationTest extends AbstractIntegrationSpec {
 
-    private static final SOME_CREDENTIALS =
-        """
-            remote.credentials {
-                username = "user"
-                password = "pass"
-            }
-        """
-
-    private static final NO_CREDENTIALS = ''
-
-    private static final INCOMPLETE_CREDENTIALS =
-        """
-            remote.credentials {
-                username = "user"
-            }
-        """
-
     @Rule
     BuildOperationsFixture buildOperations = new BuildOperationsFixture(executer, temporaryFolder)
-
-    @Rule
-    HttpBuildCache httpBuildCache = new HttpBuildCache(testDirectoryProvider)
 
     def "local build cache configuration is exposed"() {
         given:
@@ -75,79 +54,6 @@ class FinalizeBuildCacheConfigurationBuildOperationIntegrationTest extends Abstr
         result.local.push == true
 
         result.remote == null
-    }
-
-    @Unroll
-    def "remote build cache configuration is exposed"() {
-        given:
-        httpBuildCache.start()
-        def url = "${httpBuildCache.uri}/"
-        settingsFile << """
-            buildCache {  
-                local {
-                    enabled = false 
-                }
-                remote(org.gradle.caching.http.HttpBuildCache) {
-                    enabled = true 
-                    url = "$url"   
-                    push = $push 
-                    $credentials
-                }
-            }
-        """
-        executer.withBuildCacheEnabled()
-
-        when:
-        succeeds("help")
-
-        then:
-        def result = result()
-        result.enabled == true
-
-        result.remote.enabled == true
-        result.remote.className == 'org.gradle.caching.http.HttpBuildCache'
-        result.remote.config.url == url
-
-        if (authenticated) {
-            result.remote.config.authenticated == "true"
-        } else {
-            result.remote.config.authenticated == null
-        }
-
-        result.remote.type == 'HTTP'
-        result.remote.push == push
-
-        result.local.enabled == false
-
-        where:
-        authenticated | credentials            | push
-        'true'        | SOME_CREDENTIALS       | true
-        'false'       | NO_CREDENTIALS         | false
-        'false'       | INCOMPLETE_CREDENTIALS | false
-    }
-
-    def "remote build cache configuration is exposed when basic auth is encoded on the url"() {
-        given:
-        httpBuildCache.start()
-        def safeUri = httpBuildCache.uri
-        def basicAuthUri = new URI(safeUri.getScheme(), 'user@pwd', safeUri.getHost(), safeUri.getPort(), safeUri.getPath(), safeUri.getQuery(), safeUri.getFragment())
-        settingsFile << """
-            buildCache {  
-                remote(org.gradle.caching.http.HttpBuildCache) {
-                    enabled = true 
-                    url = "${basicAuthUri}/"   
-                }
-            }
-        """
-        executer.withBuildCacheEnabled()
-
-        when:
-        succeeds("help")
-
-        then:
-        def config = result().remote.config
-        config.url == safeUri.toString() + '/'
-        config.authenticated == "true"
     }
 
     def "custom build cache connector configuration is exposed"() {
@@ -222,9 +128,6 @@ class FinalizeBuildCacheConfigurationBuildOperationIntegrationTest extends Abstr
     def "disabled build cache configurations are exposed"() {
         given:
         def cacheDir = temporaryFolder.file("cache-dir").createDir()
-        httpBuildCache.start()
-        def url = "${httpBuildCache.uri}/"
-
         settingsFile << """
             buildCache {
                 local(DirectoryBuildCache) {
@@ -232,9 +135,9 @@ class FinalizeBuildCacheConfigurationBuildOperationIntegrationTest extends Abstr
                     directory = '${cacheDir.absoluteFile.toURI().toString()}'
                     push = false 
                 }
-                remote(org.gradle.caching.http.HttpBuildCache) {
+                remote(DirectoryBuildCache) {
                     enabled = false 
-                    url = "$url"   
+                    directory = '${cacheDir.absoluteFile.toURI().toString()}'   
                     push = false 
                 }
             }
@@ -249,28 +152,24 @@ class FinalizeBuildCacheConfigurationBuildOperationIntegrationTest extends Abstr
         result.enabled == false
 
         result.local.enabled == false
-        result.local.className == 'org.gradle.caching.local.DirectoryBuildCache'
+        result.local.className == DirectoryBuildCache.name
         result.local.config.location == cacheDir.absoluteFile.toString()
         result.local.type == 'directory'
         result.local.push == false
 
         result.remote.enabled == false
-        result.remote.className == 'org.gradle.caching.http.HttpBuildCache'
-        result.remote.config.url == url
-        result.remote.type == 'HTTP'
+        result.remote.className == DirectoryBuildCache.name
+        result.remote.config.location == cacheDir.absoluteFile.toString()
+        result.remote.type == 'directory'
         result.remote.push == false
     }
 
     def "remote build cache configuration is disabled when --offline is provided"() {
         given:
-        httpBuildCache.start()
-        def url = "${httpBuildCache.uri}/"
-
         settingsFile << """
             buildCache {
-                remote(org.gradle.caching.http.HttpBuildCache) {
+                remote(DirectoryBuildCache) {
                     enabled = true 
-                    url = "$url"   
                 }
             }
         """
@@ -284,7 +183,7 @@ class FinalizeBuildCacheConfigurationBuildOperationIntegrationTest extends Abstr
         result.enabled == true
 
         result.remote.enabled == false
-        result.remote.config.url == url
+        result.remote.className == DirectoryBuildCache.name
     }
 
     Map<String, ?> result() {
